Move welcome email into interactive package
This email is only for interactive users so following
path-of-least-surprise this moves it to it's, hopefully, more expected
location.
