+++ interactive/emails.py
@@ -1,4 +1,5 @@
 from django.conf import settings
+from django.urls import reverse
 from furl import furl
 from incuna_mail import send
 
@@ -18,5 +19,25 @@ def send_report_uploaded_notification(analysis_request):
         subject="Your OpenSAFELY Interactive report is ready to view",
         template_name="emails/notify_report_uploaded.txt",
         html_template_name="emails/notify_report_uploaded.html",
+        context=context,
+    )
+
+
+def send_welcome_email(user):
+    login_url = furl(settings.BASE_URL) / reverse("login")
+
+    context = {
+        "domain": settings.BASE_URL,
+        "name": user.name,
+        "url": login_url,
+    }
+
+    send(
+        to=user.email,
+        subject="Welcome to OpenSAFELY",
+        sender="notifications@jobs.opensafely.org",
+        reply_to=["OpenSAFELY Team <team@opensafely.org>"],
+        template_name="emails/welcome.txt",
+        html_template_name="emails/welcome.html",
         context=context,
     )
