+++ jobserver/emails.py
@@ -140,23 +140,3 @@ def send_token_login_used_email(user):
         template_name="emails/login_token_used.txt",
         html_template_name="emails/login_token_used.html",
     )
-
-
-def send_welcome_email(user):
-    login_url = furl(settings.BASE_URL) / reverse("login")
-
-    context = {
-        "domain": settings.BASE_URL,
-        "name": user.name,
-        "url": login_url,
-    }
-
-    send(
-        to=user.email,
-        subject="Welcome to OpenSAFELY",
-        sender="notifications@jobs.opensafely.org",
-        reply_to=["OpenSAFELY Team <team@opensafely.org>"],
-        template_name="emails/welcome.txt",
-        html_template_name="emails/welcome.html",
-        context=context,
-    )
