+++ tests/unit/interactive/test_emails.py
@@ -1,6 +1,6 @@
-from interactive.emails import send_report_uploaded_notification
+from interactive.emails import send_report_uploaded_notification, send_welcome_email
 
-from ...factories import AnalysisRequestFactory
+from ...factories import AnalysisRequestFactory, UserFactory
 
 
 def test_send_report_uploaded_notification(mailoutbox):
@@ -12,3 +12,15 @@ def test_send_report_uploaded_notification(mailoutbox):
 
     assert analysis_request.get_absolute_url() in m.body
     assert list(m.to) == [analysis_request.created_by.email]
+
+
+def test_welcome_email(mailoutbox):
+    user = UserFactory()
+
+    send_welcome_email(user)
+
+    m = mailoutbox[0]
+
+    assert user.name in m.body
+
+    assert list(m.to) == [user.email]
